apps/worker/pyproject.toml
[project]
name = "purrview-worker"
version = "0.1.0"
description = "PurrView stream processing worker - cat feeding monitor"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "google-genai>=1.0",
    "supabase>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
